Finished ListProperty, fixed linker errors, shader defines missing@
